refactor: 테스트 위치 이동 `BookTest` -> `BookRepositoryTest`
